fixing TDAR-2022 w/ Adam
Picking option 3, creating a scheduled task to check if the enabled auth system is configured properly and logging to
error if it is not.

TDAR now starts up as long as there is an enabled auth provider.  Auth providers are enabled in the spring-crowd and
spring-ldap XML configs, respectively (by importing one or the other in spring-local-settings.xml)