fixing dataset confidentiality bug, side-stepping potentially larger issue (for now)
The core issue of the dataset confidentiality bug was that it it was not copying the restriction value of the original file when calling out to processFileProxy() to create the translated version.  I solved this by properly setting this
value on the proxy prior to making the call.

I think the larger problem is that AbstractInformationResourceService#createVersion(irFile, proxy) overwrites the the irFile.restriction field even when creating a 'derivitive' version of a file.  So I'm concerned there are other opportunities to omit the restriction value in a similar fashion.  However, in practice we don't see this happening too often, and ultimately I don't have any better ideas to make createVersion()  'safer'.  Tabling the issue for now...