adding onFail()  test helper method
AbstractIntegrationTestCase will now call onFail()  after each failing test.   Default behavior is no-op, but you can override this to extra work in event of failure.  For example,  I have the extended onFail() in the web tests to log out the response html after every failing test, since this is typically what I do when I encounter a failing test.   Other ideas might be logging the contents of a filestore node or copying a resource xml to the surefiire directory.