removing ID from eq/hC scheme  (following the  'business key' approach down the line).
Resource currently defines equality to be based exclusively on ID.   I'm not so sure this is a "bad" thing though.    By default,  Java defines Object equality the same as Object identity, and you can use objects in sets and lists so long as you are cool with this definition and don't need a more robust definition of value  for when you put stuff in lists and sets.   What we're doing in this commit is similar:  we have a base definition of equality that is based off of *database* identity and we override this definition when we need more nuanced definitions of value when we compare objects and put them in collections.