fixing latent bug in ResourceCollectionService, @Ignoring broken test.
resourceCollectionService was doing this:
  current.add(collectionToAdd);
  collectionToAdd.getResources().add(resource);
  resource.getResourceCollections().add(collectionToAdd);

The problem is,  current and resource.getResourceCollections() are the same object, so add() is called twice.   This wouldn't be a problem because a set should weed out the dupe, but when you call
    collectionToAdd.getResources().add(resource);

you have changed the hashcode of collectionToAdd and corrupted the set. so you wind up with a set that contains two identical items.