adding test  that explores one way of having consistent definition of identity with a contextual definition of "similarity".
TreeSet allows you to supply a comparator which it will use in place of equality (primarily for sorting purposes).   We can use this, for example,  to filter items we consider items to be "similar"  even though they are not identitcal.    Note that the java documentation *discourages* using comparitors that are not consistent with hashcode and equals(), but it is not prohibited, and it might be possible to mitigate the risks of doing so.

For example:  it would clearly be bad  we replaced all of a Resources set getters with TreeSets that used a goofy comparator.   However,   if we wanted to flatten  an incoming list of Person objects we could,  withn the scope of a method designed to de-dupe similar records, create a TreeSet with a 'is similar to'   comparator, place the filtered results in an ArrayList, and discard the treeset prior before the method returns.

I don't know if I like this particular approach, but I'd be willing to abandon "business keys everywhere!"  if we could have an easy way to define "similarity"  (and, even better, if we could define similarity in multiple ways to accomodate various contexts).