the searchbox returns
I've managed to get the searchbox to appear,  though the html is invalid (you can't put a form in a LI) and it relies on yet one more obtuse css rule for correct appearance.    The header needs much refactoring.

For example,  the search box only fits underneath the 'welcome back' div because we've set the margin so high that it appears below the floating 'welcome back' section.    However, because the margin increases the element height we must also increase line-height so that the "BROWSE" and "UPLOAD" button labels are centered correctly.   The header is filled with dozens of these interdependencies such that a slight change to one element has adverse effects on another.