fixing bug in Persistable.isEqual()
Our equality test was only checking to see if A was assinable from B,  so you could wind up with a situation where a.equals(b) == true but b.equals(a) == false.  So what we need to do is check if two objects are in the same class heirarchy   (a subclass of b OR b subclass of a).