adding scan metadata templates & reminder div (needs better lingo)
In this update, the current idea is that there will be one async file upload section that accepts images, zips, and xls/csv .  When you select a "scanner technology" type,  the "template reminder" div appears that informs the user that they can upload scan metadata using one of the provided tamplate samples.

Currently these are all "suggestions" to the user  and it does not hook into form validation.   We can easily add hooks so that an xls is required depending on scanType value, but this adds a dependency to the tdar.fileuploadvalidtation plugin.