more work on FileUploadITCase
It's still broken but I think I'm on the right track.  The current error is "element is not currently visible and so may not be interacted with".  I modified the 'opacity', and while it altered the display (you can see a small portion of the 'real' file input element)  I still received the same selenium error.  I think we need to remove undo all of the styles applied to that element to be sure.  Such as:
	opacity
	filter
	zorder
	position
	left, top