making  some progress on FileUploadITCase by disabling CSS.
The test is now getting a little further along, and the fileupload widget appears to fire up and show the "sending file" UI.  However, the upload does not complete.  Either I'm not giving it enough time or there is some other javascript error taking place (perhas, ironically, because CSS is turned off).

TODO:  instead of turning off CSS entirely,  just override all of the styles on the fileinput that make selenium think that it is not visible.
TODO:  redirect javascript console output to logger.