trying to suss out cause of javascript errors on bamboo tests.
we are getting test failures on bamboo because reportJavascriptErrors finds js errors, but this only happpens on the build box,  not on my local instance.   Furthermore,  the argument passed to window.onerror is an event, instead of the error message string.  So if that happens we're dumping the contents of the event into __errorMessages in hopes that it will give us more evidence.