tweaking geotiff selenium test to see if previous failure was timing issue
I think the geotiff selenium test was (occasionally) failing because it tries to fill in the login page before it appears.  Selenium is supposed to block  on click() when the click would cause navigation, but I notice that this does not always occur (perhaps selenium only blocks after the browser's onclick is finished, and in that time our selenium test has already tried to fill out the loginUsername element?).  Either way, I put a waitFor()  for the loginUsername element, and that seemed to make the test consistently pass on my end.