more work on tos workflow.
most things are wired up, but doesn't work quite right yet.  what happens is that the user accepts the agreements, the acceptAcion updates the user record and indicates that all requirements have been met and returns SUCCESS.  however,  when the client redirects to the dashboard the user record seems to have not been updated yet, and so the system redirects back to /show-notices.  even weirder,  show-notices is empty because it detects that the user tos/contributorAgreement fields *have* been updated.  weird.