finally got basic user-agreement workflow working, but it feels like a kludge
I updated the user-agreement workflow:
1) the actions now run in the 'secured' stack, and I updated intercept() to watch out for infinite redirect
2) intercept() now does a refresh on the sessionData.authenticatedUser.

I think that 2) is what is really making the process finally work.   Without the change,   I have this weird scenario where  I accept all the user-agreements, the controller detects that all agreements are satisfied and redirects to /dashboard,  but the intercept() of the /dashboard request detects that the agreements *aren't* satisfied, so it redirects back to /show-notices.

I think the problem is that the succesful /agreement-response request  and the /dashboard request are serviced by different threads,   and so in the /dashboard thread the sessionData doesn't have the updated user record.   I think that's why refresh() fixes the problem, but I don't like that solution because it introduces an extra SQL lookup in almost every request on  on the authenticated stack.