adding quotes to css attribute selectors (potentially fixes TDAR-3390)
Chrome seems to be reporting css attribute-selector syntax errors that it hadn't been reporting in the past (or maybe I've just not noticed them?). Apparently attribute selector values need to be quoted, so I've fixed a few files where we use them, including jquery.maphighlight extension.  Maphighlight isn't one of ours, but we've already modified it so I didn't see the harm (I did put a note at the top  of the file).