disregard some gis validation rules if we think user is uploading raster/map files (as opposed to shape files)
I don't think this is a good long-term approach.   These rules are built on the assumption that the user is uploading gis shapefiles, as opposed to a gis raster files or gis map files (which may also include a raster or shape files).   If we do allow these other gis resource types, I think we need another form field where we ask the user what gis resource type they are uploading.  Then use a different set of validation rules for each of the three types.