work-around for issue w/ jquery-ui and bootstrap 2.3
We rely on a recent point release of jQuery-UI for our "autocomplete cache" feature, but it also introduces an issue w/ bootstrap form fields w/ column-sizing.
Specifically, $UI caches autocomplete menu-items as hidden elements adjacent to the autocomplete input field.  This breaks a "rule" in bootstrap 2.3 where direct
children of a .row element must have column sizing. To get around this, we need to wrap the autocomplete field in a column-sized elemeng (e.g.
<div class="span3">)