if async upload fails, show error then remove row from files table (fixes TDAR-3574)
smaller bug: if upload fails while doing a file replace,  the client still removes the broken row after a timeout
but does not display an error message.  I'm still looking into this, but fortunately the original file is not
adversely affected and the client is not put in some weird state; in other words, the user can ostensibly retry
the replace operation.