enabling firefox profile: "focusmanager.testmode"  (WARNING: THIS MAY BREAK UNIX SELENIUM TESTS)
Firefox OSX does not emit UI events (focus, blur, change, etc)  when the application is not in the foreground+active.  This is correct behavior but can make it a pain to test on osx.  Enabling "focusmanager.testmode"  seems to solve this problem for me, but there are reports that it totally hoses Firefox for Unix.   So this might break our selenium tests in bamboo.