workaround for replace-file bug in IE9
We disable the replace "button" when we are in the middle of uploading a replacement file so that we don't wind up in this weird state where the user clicks the replace button in the middle of a long-running upload.  This breaks ie9 replacement uploads, because IE9 uses iframes to upload the file in a separate form  (and disabled form inputs are not included in a POST request).