registration fix for bug where user causes PK violation with a well-timed double-submit (TDAR-3777)
If a user double-submits the registration form, it can result in two threads that service two separate registration requests.  Since the requests are close-enough together,   thread#2 does not find a pre-existing user (because thread#1 hasn't created it yet), so it proceeds to create a user with the same username as thread#1.  So thread#2 fails, rolls back, and renders an error page to the user even though thread#1 has successfully created the account.

I've attempted to prevent the issue on the client and server. On the client we disable the submit button after the first submit.   On the server side we synchronize the addAndAuthenticate() method to prevent the dirty read that may stem from a double-submit.  We could elevate the transaction isolation level, but this could have significant performance implications.