reformatting: ftl, js, ftl, xml  (see notes)
There should only be formatting changes in this commit. If you find any effective/functional changes consider it a bug that requires a ticket.

For javascript files, I've attempted to adhere to the settings specified in tdarEclipseJavascriptFormat.  One exception is that I am enforcing
braces for if-statements.  This has the potential to introduce functional changes -- I've reviewed where these replacements occur and I think
we're safe.

For ftl,html, and css files, there are a minimal set of rules applied:
- replace tabs w/ spaces
- tab size: 4
- continuation size: 8
- line endings: LF
- insert LF at EOF: true
- hex colors: lowercase (css only)

Note:  use @formatter hints to retain original formatting in a file or part of a file, e.g:

    /* @formatter:off  */