pulling out integrationColumn cleanup into separate method (TDAR-3852)
when struts builds an object graph on the controller it potentially introduces null items into the
controller's collection fields (e.g. when struts is trying to populate mylist[1] but mylist[0] does not yet
exist, struts will call controller. getMylist().add(null)  prior to setting the value for mylist[1].)

In our controller we were aggressively cleaning up the integration columns by putting cleaning in the getter
method.  This caused caused a bug when struts tried to build the controller object graph following a
submit where the user had deleted columns or sub-columns from the middle of the list in the select-columns UI.

The fix was to remove the active cleanup and perform this cleanup step in the start of the controller execute
phase.