merge from stable (see notes)
WorkspaceController differs slightly in that we clean up the integrationColumn lists in prepare() as opposed to the
filter action.  I was going to introduce this change in a separate commit but I already had a merge conflict
in WorkspaceController anyway so I went ahead and added the aforementioned logic change as well.