Refactoring SheetEvaluator and ExcelConverter for TDAR-3630.
SheetEvaluator now does more sophisticated bookkeeping to determine a given Sheet's header and data row and column
bounding indices. Headers are loosely defined as rows with the following properties:

1. the same quantity of data as the rest of the rows being evaluated
2. mostly alphabetic data (> 70% a-zA-Z characters)
3. sane average column name length of < 30 characters
4. no more than 2 blank characters (this is arbitrary, should be revisited)
5. in the first 10 rows of the sheet

(TODO: revisit / refactor these assumptions if they do not hold in practice)

A few of the tests and the way ExcelConverter generates DataTableColumns have been changed. For degenerate excel files
(testMalformedExcelDatasetFromScott) I'd argue that we shouldn't even preserve non-tabular data sheets (the trick is
figuring out what "tabular" data is, I have a stubbed out method for the moment that will rarely trigger) since there is
no meaningful data integration that could be done with it. The main change for now though is that the ExcelConverter
only generates data table columns for columns in the Sheet that contain data. So if tabular data has offsets (ranging
from E7..J7 for example) there would only be 6 columns created instead of 10 columns created.