Temporary solution: Adding @WriteableSession to /cart/process-choice, removing  call to  markWritable()
invoiceService.processInvoice() works with transient invoices but was throwing an exception when saving persisted invoice.  The culprit seems to be the call to markWritable(invoice).  markWritable sets the Session.readOnly = false and evicts the object from the session.  In spite of this,  attempting to save an invoice after calling markWriteable(invoice) still fails with "a different object with the same identifier value was already associated with the session".  I wonder if the problem is due to the invoice being in the second-level cache (session.evict() only removes items from the first-level cache).

For the record: I don't like the idea of writeable sessions for unauthenticated actions.  I think a better approach might be to store a transient invoice in the httpSession until the user advances past the authentication stage.