pulling out registration action processing into a service (unfinished and largely untested!!)
The motivation for having a "controller service"  came from the fact that I was writing a struts action for handling user registration in a purchasing context.  If I subclass AbstractCartAction,  I need to add getter, validation, and execute methods that are largely duplicative of what are already in AbstractRegistrationController.   The reverse is also true.   What do?

My solution thus far is to have a service class that  can validate/process a RegistrationProvider, and (potentially) add action errors to that controller should validation and/or processing experience any problems.