adding test for nelnet 'failure' response (need more realistic failures)
this test seems pointless, but its helpful in that it causes an exception that was previously getting handled by the default struts exception handler.  This is bad for the nelnet endpoint because nelnet literally expects an http response of 'success' or 'failure',  not the default tdar error message html.

Still,  we need a better example to test. For example,  an invoice that is not modifiable,  or a corrupt querystring
(e.g. hashed parameters do not match checksum from nelnet).