fixed issue where new invoice notification was not being sent (see notes for question)
There's also a call to sendSuccessEmail(), however, my understanding is that nelnet sends us notifications for successful and unsuccessful transactions (e.g. NSF, cancelled transaction, etc.). So I've renamed the method.  Also  the docs imply that the user gets an email but I only see an email going to the billing admin.  I've updated the javadoc but let me know if I'm mistaken.