adding fixme for billing account choice hack
The only way that a user ever gets assigned a default billing account is in the 'validate' method of the 'process-billing-account-choice' action. There are a couple
of things wrong with this:

1) the only expected side effects of something named 'validate' should be action errors. Instead we're doing no validation (it's commented out)  and creating a default
billing account.

2) Ideally the user should skip the redundant 'review' page if they have authenticated and they do not have any billing accounts.  In this situation the
login/registration controller would tell the service layer to create the implicit billing account (and assign it to the invoice) and then redirect the user to
the 'process-payment' page.