backout of contributor agreement changes (d19c9a1879f6, 434649ed0964, 837d9d6724f1)
turns out that contributor acceptance enforcement was working, but that the 'version number' of the agreement was set to zero on alpha (and not set in default tdar.properties, so users would not be asked to accept the contributor agreement following next login.  I've set the proper values in tdar.properties on alpha.tdar.org,  and I've backed out the commits
that altered the contributor agreement behavior (5830, 5827, and part of 5826).