make sure non-contributor eventually accepts contributor agreement after making a purchase  (see notes)
previously we were just setting the contributor flag after user selects the billing account.  However, the problem with this is that the TOS interstitial   redirects the user to the dashboard upon acceptance (thus aborting the purchase process).   The other approach (which I've taken in this commit) is to modify the authentication interceptor to simply not check for pending acceptance requirements when servicing anything in the /cart namespace.  However, for this approach to worh the authentication interceptor needs to invoke *after* the action mapping interceptor (otherwise we wont yet  know what the namespace is when the auth interceptor invokes).  So I modified the struts.xml accordingly.  this approach appears to work but I'm not sure what the other consequences may be.