incomplete fix for bulk upload issue (TDAR-4009)
The asyncStatusMap in the bulkupload service was using a weak hashmap.  I assume that either the keys are getting reclaimed prior to the lookup in checkstatus,  or that the identity-based-equality lookup for the  key(ticketId) fails because the the weakhashmap uses identity equality for looking up keys (instead of .equals()) and  the ticketId identityHashcode  changes from one request to the next  (only number objects  between  -127 and  127 will have  identity equality,    for everything else you need to use .equals().).

more info:  http://stackoverflow.com/questions/2602636/why-cant-the-compiler-jvm-just-make-autoboxing-just-work

Reverting to a normal (strong reference) concurrent hashmap seems to work, but I'm guessing if we stick with this solution we will need to be more vigilent about pruning the map after  the bulkuploads finish (either in success or failure).