replacing inline java in freemarker expressions, where possible.
Some parts are tricky and are best left as java expressions.  For example,  you can sometimes replace ${collection.isEmpty()} with ${collection.empty}, but depending upon the state of the  collection and/or whether it is a List vs. a Set, freemarker may wrap the object in such a way that ".empty" is a syntax error.  So it's best to rewrite in the form of ${collection?has_content}  or leave it as a java expression.

Also tricky is non-string hashtable keys.   Freemarker hashes implicitly have string keys, and so the behavior with object keys is undefined (though likely will not work).   If you expose a hashmap  / linkedhashmap as <string, object> you'll probably be fine.