fixing checkbox data-binding
Rather than designating observables explicitly+programmatically, Allen recommends simply using the knockout mapping plugin to have to perform this mapping automagically.  This is ostensibly less efficient (though it doesn't appear to be noticeable in the current demo), the advantage is that you end up interacting with the viewmodel in a much more consistent way. Plus,  to facilitate certain behaviors you may need to map a lot of two-way binds, and that could end up being quite tedious.