adding initial hosted-download test (expected to fail)
This is a simple test that requests a hosted-download url from an unauthenticated client. When complete, the test should set up the necessary permissions such that downloads
are successful when the apikey+referrer+fileid are correct and fail if either the apikey/referer is incorrect or the fileversionid is not in the accepted list.