fleshing out edit-angular
This was a little easier-going.  I put all of the logic in a single controller, however, in angular-world it seems more common that controllers are associated with sections within a page.  So for example, it might make more sense to have a separate controller that handles the state of the tabstrip.  This would require that the integration data be shared between controllers --  angular advocates using 'services' for this, but I'm not sure I want to do that just yet (I could put the data in global scope or a shared closure, but that feels like cheating)