set search filter param names to comply with the struts endpoint
This is still dirty, we need to map the javascript view model object properties and add a struts prefix to it so that it
matches what Struts wants for form input names. Marginally better solution would be to introspect on the JS properties
and dynamically construct a Struts-compliant map from the list of desired properties with the appropriate prefix.