more shellgame antics: DocumentDataFactory is now DataService
Currently the code for loading/translating data  (embedded data, ajax data) is spread across multiple files.   It makes more sense to centralize this access into a single place.   This will be especially helpful when we actually wire up  saving/loading  previous integrations.

Furthermore,  I think it's a good idea to use DataService.js to describe the 'schema' of the data it expects from the server.   At the very least describe in comments, but ideally it might be good to use mapping functions for each ajax url  (e.g.  translateIntegrationColumnDatails(rawData) )   That way,  if we decide to make
changes to the schema of the ajax server endpoints  it will be easier to modify the affected js code.