adding data-integration/tests.js  (see notes)
This file has some ideas on functional tests (that should be fleshed out and placed in confluence)   as well as a minimal implementation of one way we *could* write some barebones test.

The way it works is that you would uncomment the script include,  and the file would register itself  as an addon to integrationApp.   The file doesn't add any components or expose any functionality.  All it does is includes a example "run" functions.  Angular executes these functions after it has loaded all of your apps resources (controllers, services, templates, etc).