cart: show contributor-agreement checkbox on review page if authenticated user is not a contributor
If user does not check the box,  they are redirected back  /cart/add to start again.   I would prefer to just have CartBillingAccountController render review.ftl, however,  but it does not have the same properties and CartReviewPurchaseAction and so you wind up with an ftl error.   Still, this is a much better workflow that what we had before (user gets redirected to accept-user-agreement and then lands on dashboard, effectively cancelling the purchase)