don't add action error for status code 403 (FORBIDDEN)
My initial thought was that we should simply remove this from the 403.ftl template, but since this uses the same layout-bootstrap template, this will be a pain to
do and I am loathe to add another if/then to that template.

Fortunately, it seems that we only add this error message for fall-through case of checkValidRequest().  So this seems like a safe place to remove the call to addActionError().