temporary: adding static content servlet for whitelabel files
It'd be nice to serve static content from someplace other than the filestore or the webapp folder; for content that does not require the infrastructure & gaurantees
provided by the filestore, but also not bundled as part of the application jar (e.g. whitelabel logos, ads, etc.)


For now I made a dumb servlet that does this, but it is not a good long-term solution. Ostensibly static content should be served by a dedicated http server (apache,
nginx), or perhaps by the default servlet of the app container (e.g. Tomcat's  DefaultServlet).