workaround for firefox autocomplete bug (TDAR-4670)
I think this workaround essentially duplicates the purpose of the "autoVal"  attribute (i.e. save the value of the most recent menu item selection),  however I was not able to get it to work, and I suspect that it either unused and/or broken.

I think that our approach for autocomplete is ripe for a refactor so I'm regarding this as a workaround as opposed to a straight-up "fix".