karma: using wro.xml to build config.files array
I created a node util to build a list of js/css files that we can use in the karma.config.files array.  I'm not happy with this solution because it violates the rule of least-surprise, but it's slightly better than manually keeping the karma.config file in sync with the wro.xml file.   Hopefully we can move to a simpler solution (e.g. switth to grunt for minification, or seeing if wro supports a .json config format).

You'll notice that the tests still fail.  Hopefully this is due to the fact that they are not implemented.   In other words,  I think that the karma config is finally
including the same js files (in the same order) as the that get included in a typical tDAR page.

todo:
- include css files
- possibly create second karma profile for ng-integrate.  Besides the different includes, we will want to load angular-specific plugins.