karma: adding helper for mocking JSON-P responses
Jasmine can mock json ajax responseTypes, but it doesn't support jsonp because it has no way of knowing the name of the calling function that wraps the json.    So, I wrote a helper that expects the jquery convention of storing the function name in a 'callback' parameter (either in POST data or GET params).