adding init calls to specs; maybe found karma ajax issue (see notes)
The pricing spec test calls init(),  and that is the reason behind the karma 404 warning  (init() makes  ajax call to localhost, karma responds w/ 404 because the url is not mapped).

The problem is that when I try to install  jasmine.Ajax(), the calls to loadFixtures() no longer work.   This is a known issue by the jasmine-jquery team, however, it doesn't always affect us like it does in PricingSpec.  For example,   BulkSpec works just fine even though it also loads the jasmine.Ajax() interceptor prior to calling loadFixtures().

if you reverse the order,  (loadFixtures() prior to installing jasmine.Ajax())  the interceptors don't properly register.

Potential work-arounds:
-   for ajax-related tests,  manually build dom w/ setFixtures()  instead of loading them dynamically.   This *sucks* for big fixtures.
-  use the other fixture mechanism  which loads strings in a javascript array.  Not a fan of this either because it will load the fixtures in ram for the entirety of the time karma is running.