adding test for pdf merger bug (see notes)  (TDAR-4906)
Note that this test does not include the files necessary to recreate the bug (hence the test is @Ignore'd).   You set up the test by copying a coverpage (look in /tmp) and the document mentioned in TDAR-4906.  Copy these files as "coverpage.pdf" and "document.pdf" to the /tmp folder (or modify the TMP_FOLDER variable.