fixed: repeated upload attempt always fails if first attempt fails (TDAR-5061)
Previously, If a file transfer failed in the course of  a file replacement, the client would not respond if a user attempted to upload the same file again.  This is because the underlying file input  value attribute did not change,  and thus the browser did not fire the "change" event (because nothing changed).    To work around this problem, the error handler will now clear the value attribute. This has the effect of forcing any subsequent click on the upload button to trigger a change event (unless the user cancels the upload dialog, which is fine).

Note that users still cannot upload the same file multiple times if the first attempt was successful.