Forward malformed keyword requests to 400/Bad Request  (fixes: https://issues.tdar.org/browse/TDAR-5037).
We've configured struts to retain actionErrors for redirect responses.   As a result,  struts appends these actionErrrors to the action on the receiving-end of the redirect.  In this instance,    the BrowseKeywordController received bad input, and correctly added action errors to the actionErrors list and returned INPUT, which is mapped to a redirect to "/not-found"".   In the subsequent request for "/not-found", struts appended the actionErrors from the previous action to the current action,  thus causing the validation interceptor to set the action result to INPUT.  Because the /not-found action has no mapping for INPUT, struts threw the "no mapping" exception.

The fix is to simply render the error.ftl with a 400 status code (bad request), instead of performing a redirect.  The client still receives the pertinent error information without the aforementioned side-effect.