attempt to fix broken date serialization (see notes)
I tried changing the json serialization format for dates so that it emits dates a "February 19, 2016" instead of "2016-02-16" (due to a peculiarity in Javascript,  dates written in the latter syntax are parsed as GMT whereas the former are parsed as the local time zone).  I modified the JacksonUtils initializeObjectyMapper() to use this new date format, however,  the webapp still seems to be serializing using the "YYYY-MM-dd" format.

I wrote a unit test to assert the correct behavior, and the test passes (both in IDE and via maven ), however, when the webapp renders (for example) an Image edit page,  it serializes the  "filesJson" property using this YYYY-MM-dd format.  I'm not sure what I'm doing wrong here.