shrinkwrap, take 3
The information that I've read indicates that you cn instruct npm to ignore optional, platform-specific dependencies.
However, I can't figure out how to make that happen.

So now I'm taking another approach: generate the shrinkwrap file on Linux and check that in instead.  This seems to
work on build, and does not create the same-but-opposite install problems when I try to perform npm-install
on my Mac (haven't tried Windows yet).

This is not a good long-term solution, but hopefully buys us some time.