Fixing more tests - adding workaround for htmlunit tests that don't expect more than one submittable form per page.
Our htmlunit base classes could really benefit from a refactor.   In the selenium tests it's easy to identify the specific form + button that you wish to click, but the htmlunit abstract class methods tend to assume one form (or, if more than one form,  only one form that you plan on interacting with).  I put in some quick fixes for a couple of the more stubborn tests, but it'd be nice to have something similar to  the find() + WebElementSelection api we have for the selenium tests.