removing extraneous RegexFilter (also, here are those notes I promissed)
RegexFilter, by default,  does not enable the MULTILINE and DOTALL flags.  Without those flags enabled you can't write patterns that match anytihng that spans lines.   There is supposedly a way to pass flags to the RegexFilter constructor, but it is not documented, and I haven't guessed the correct magic words to get it to work.   In the meantime, I've changed our javascript error reporting so that we strip carriage-returns from  the client error messages.