removing sketchy assertions from CommonSpec
One of the tests was asserting that window.setTimeout is not called, which is a popular function invoked all over the place, so there's a small chance of another caller firing it  in the time this test runs.

Better to check $.fn.dialog instead.   Even *better* would be to refactor tdar.common.sesssionTimeoutWarning() someday.