adding jquery validation rule "equalToIgnoreCase" (see notes)
Technically, email addresses are case-sensitive though most providers treat it as case-insensitive - and our tests expect this behavior as well.  The reason why the equalTo rule on confirmEmail started breaking our tests is because it was never really enforced in the first place -- the configuration contained a typo  and referred to a non-existent fileld named "confirmEmail" instead of "registration.email".