fixing minor display issue with button group
Bootstrap has some css classes vary their style based upon the ordinal value of an element.  In this instance, the last button in a "button group" should have a tapered right edge.  However,  the hidden input fields after the last button meant that the css rule was not being applied.