omit empty facet option parameters for empty lists (fixes TDAR-5574)
The facet links specify  parameters for fileAccess, documentType, and integratableOptions  even when those lists are empty (e.g.   ""?integratableOptions=" without any value).  This is unnecessary as the action class initializes them to be empty.  And in Struts 2.5.5,   the parameter mapper interprets this as a list with a null element,  which is not what we want.