attempting to abide by isListCollectionsEnabled() (cheating where I can't)
I'm trying to incorporate the listCollectionEnabledFlag, but in some places it's *extremely* difficult.  Case in point: OrganizeAction.  The complexity of the generic type parameters is such that is near-impossible to instantiate and manipulate lists that contain either ListCollections or SharedCollections.   When you find yourself writing declarations like this:

    List<? extends HierarchicalCollection> resourceCollections = TdarConfiguration.getInstance().isListCollectionsEnabled() ? new ArrayList<ListCollection<ListCollection>> : ArrayList<SharedCollection<SharedCollection>>;

It is time to seriously reconsider the viability of committing to this flag-based approach.