Clicking the buttons to remove from managed/unmanaged add to the vue model. Theres a link to undo the modification. Removing the modification will remove from vue, and should cause the datatable to update to re-enable the button.
The datatable is not maintaining state of the $datatable.data() properties that were set. I am not sure if $datatables.fnDraw() causes it to reinitalize the entire data values, or if the values were not being set correctly to begin. In either case, the result is that when the page is redrawn,
all buttons are re-enabled, including those that should not be.