The obfuscation needs to take into account the following:
X = width of the bounding box ( <=1 mi)
Z = Width of the  obfuscating box (>=1 mi)
Y = The delta of Z-X (the difference of how big one box is from the other).

X' = 1/2 the width of the bounding box , the minimum amount of space needed so that obsfucating box covers the edge.
Y' = The delta of Y-X'  - How much space the bounding box can be shifted back so that the edge is still covered.

In the cases where Y' > X', that means that the box can be shifted up to random() * mile, and the shift occurs ok. The obsfucating box is sufficently large enough that the random() coefficent can operate properly without causing problems.

In the case Y' < X', it means the boxes are in similar size, but an increasing random() coefficent causes the box to shift too far right because there isn't enough buffer between the two to allow the shift to happen. To prevent this, the shift limit is capped at X'.